+++ Modules/Tours/MITToursTour.m
@@ -60,8 +60,8 @@ static NSString *const kMITToursSideTrip = @"Side Trip";
     NSInteger hours = [self.estimatedDurationInMinutes integerValue] / 60;
     NSInteger minutes = [self.estimatedDurationInMinutes integerValue] % 60;
     
-    NSString *hoursString = (hours > 0) ? [NSString stringWithFormat:@"%d hour%@ ", hours, (hours != 1) ? @"s" : @""] : @"";
-    NSString *minutesString = (minutes > 0) ? [NSString stringWithFormat:@"%@%d minute%@", hoursString.length > 0 ? @"and " : @"", minutes, (minutes != 1) ? @"s" : @""] : @"";
+    NSString *hoursString = (hours > 0) ? [NSString stringWithFormat:@"%ld hour%@ ", (long)hours, (hours != 1) ? @"s" : @""] : @"";
+    NSString *minutesString = (minutes > 0) ? [NSString stringWithFormat:@"%@%ld minute%@", hoursString.length > 0 ? @"and " : @"", (long)minutes, (minutes != 1) ? @"s" : @""] : @"";
 
     return [hoursString stringByAppendingString:minutesString];
 }
